+++ Server/src/meplay/evpp/udp_server.cc
@@ -221,7 +221,7 @@ void Server::RecvingLoop(RecvThread* thread) {
 
         MessagePtr recv_msg(new Message(thread->fd(), recv_buf_size_));
         socklen_t addr_len = sizeof(struct sockaddr);
-        int readn = ::recvfrom(thread->fd(), (char*)recv_msg->WriteBegin(), recv_buf_size_, 0, recv_msg->mutable_remote_addr(), &addr_len);
+        int readn = ::recvfrom(thread->fd(), (char*)recv_msg->WriteBegin(), static_cast<int>(recv_buf_size_), 0, recv_msg->mutable_remote_addr(), &addr_len);
         if (readn >= 0) {
             MPLOG_TRACE << "fd=" << thread->fd() << " port=" << thread->port()
                       << " recv len=" << readn << " from " << sock::ToIPPort(recv_msg->remote_addr());
